Enable VM to edit resolv.conf per Issue #198
This adds become: yes to the play run on localhost to edit resolv.conf.
On a container this works without become but on a VM this requires sudo access.
Enabling sudo on a container should not have a negative affect. This can
also be accomplished by setting the variable ansible_sudo: true but that
is a per-role or per-task setting, where become is per-play.